+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Release 1.0.1
 
 * [Refactor]: removed public router() method from App class
+* [Refactor]: refactored app() function
 
 ## Core Release