+++ CHANGELOG.md
@@ -1,10 +1,11 @@
 # Changelog
 
-## Release 1.0.1
+## Release 1.1.0
 
 * **[Feat]**: added support to disable Gutenberg Editor
 * **[Refactor]**: removed public router() method from App class
 * **[Refactor]**: refactored app() function
+* **[Fix]**: improved model stub for Post Type and Taxonomy
 
 Release Date: August 17, 2021